member2/requirements.txt
langchain
openai
pandas
numpy
faiss-cpu
python-dotenv
